requirements.txt
dearpygui>=1.3.0
torch
torchvision
pillow

